Separate GenotypeArray encoding functions into a mixin
