+++ pandas_genomics/arrays/genotype_array.py
@@ -8,6 +8,7 @@ from pandas.core.arrays import ExtensionArray, BooleanArray, IntegerArray
 from pandas.core.dtypes.dtypes import register_extension_dtype, PandasExtensionDtype
 from pandas.core.dtypes.inference import is_list_like
 
+from pandas_genomics.arrays.encoding_mixin import EncodingMixin
 from pandas_genomics.scalars import Variant, Genotype, MISSING_IDX
 
 
@@ -232,7 +233,7 @@ class GenotypeDtype(PandasExtensionDtype):
             )
 
 
-class GenotypeArray(ExtensionArray):
+class GenotypeArray(ExtensionArray, EncodingMixin):
     """
     Holder for genotypes
 
@@ -720,9 +721,9 @@ class GenotypeArray(ExtensionArray):
             return NotImplemented
         return (self.allele_idxs >= allele_idxs).all(axis=1)
 
-    #####################
-    # Utility Functions #
-    #####################
+    ###################
+    # Utility Methods #
+    ###################
     def set_reference(self, allele: Union[str, int]) -> None:
         """
         Change the reference allele (in-place) by specifying an allele index value or an allele string
@@ -774,94 +775,3 @@ class GenotypeArray(ExtensionArray):
         self._data["allele_idxs"][was_ref] = allele_idx
         # What was the allele is now reference (0)
         self._data["allele_idxs"][was_allele] = 0
-
-    ######################
-    # Encoding Functions #
-    ######################
-
-    def encode_additive(self) -> pd.arrays.IntegerArray:
-        """
-        Returns
-        -------
-        pd.arrays.IntegerArray
-            Number of copies of the minor allele
-            pd.NA when any alleles are missing
-            Raises ValueError if there is more than 1 alternate allele
-        """
-        # TODO: Return multiple arrays for multiple alternate alleles?
-        if len(self.variant.alleles) > 2:
-            raise ValueError("Additive encoding can only be used with one allele")
-
-        allele_sum = self.allele_idxs.sum(axis=1).astype("float")
-        allele_sum[(self.allele_idxs == MISSING_IDX).any(axis=1)] = np.nan
-        result = pd.array(data=allele_sum, dtype="UInt8")
-        return result
-
-    def encode_dominant(self) -> pd.arrays.IntegerArray:
-        """
-        Returns
-        -------
-        pd.arrays.IntegerArray
-            0 for no copies of the minor allele
-            1 for any copies of the minor allele
-            pd.NA when any alleles are missing
-            Raises an error if there is more than 1 alternate allele
-        """
-        # TODO: Return multiple arrays for multiple alternate alleles?
-        if len(self.variant.alleles) > 2:
-            raise ValueError("Dominant encoding can only be used with one allele")
-
-        has_minor = (self.allele_idxs == 1).any(axis=1).astype("float")
-        has_minor[(self.allele_idxs == MISSING_IDX).any(axis=1)] = np.nan
-        result = pd.array(data=has_minor, dtype="UInt8")
-        return result
-
-    def encode_recessive(self) -> pd.arrays.IntegerArray:
-        """
-        Returns
-        -------
-        pd.arrays.IntegerArray
-            1 for Homozygous Alt
-            0 for anything else
-            pd.NA when any alleles are missing
-            Raises an error if there is more than 1 alternate allele
-        """
-        # TODO: Return multiple arrays for multiple alternate alleles?
-        if len(self.variant.alleles) > 2:
-            raise ValueError("Recessive encoding can only be used with one allele")
-
-        all_minor = (self.allele_idxs == 1).all(axis=1).astype("float")
-        all_minor[(self.allele_idxs == MISSING_IDX).any(axis=1)] = np.nan
-        result = pd.array(data=all_minor, dtype="UInt8")
-        return result
-
-    def encode_codominant(self) -> pd.arrays.Categorical:
-        """
-        This encodes the genotype into three categories.  When utilized in regression, this results in two variables
-        due to dummy encoding- "Het" as 0 or 1 and "Hom" as 0 or 1.  0 in both indicates "Ref".
-
-        Returns
-        -------
-        pd.arrays.Categorical
-            'Ref' for Homozygous Reference
-            'Het' for Heterozygous
-            'Hom' for Homozygous Alt
-            pd.NA for missing
-            Raises an error if there is more than 1 alternate allele or ploidy is not 2
-        """
-        # TODO: Return multiple arrays for multiple alternate alleles?
-        if len(self.variant.alleles) > 2:
-            raise ValueError("Codominant encoding can only be used with one allele")
-        if self.dtype.variant.ploidy != 2:
-            raise ValueError(
-                "Codominant encoding can only be used with diploid genotypes"
-            )
-
-        allele_sum = self.allele_idxs.sum(axis=1)
-        categories = ["Ref", "Het", "Hom"]
-        result = pd.Categorical(
-            values=[categories[n] if n in {0, 1, 2} else None for n in allele_sum],
-            categories=categories,
-            ordered=True,
-        )
-        return result
